+++ nomination/templatetags/nomination_tags.py
@@ -1,5 +1,8 @@
 from django import template
 from django.template.defaultfilters import stringfilter
+from django.utils.safestring import mark_safe
+
+import markdown
 
 
 register = template.Library()
@@ -14,3 +17,9 @@ def replaceunderscores(value):
 @register.filter
 def get_item(dictionary, key):
     return dictionary.get(key)
+
+
+@register.filter
+@stringfilter
+def render_markdown(value):
+    return mark_safe(markdown.markdown(value, output_format='html5'))